use float result

+++ tools/ci_tools/check_benchmark_result.py
@@ -14,7 +14,7 @@ def main():
         last_line = f.readlines()[-1]
     print()
     print("last line from log:", last_line)
-    result = int(last_line.split(' ')[-1])
+    result = float(last_line.split(' ')[-1])
     print("average time:", result)
     print("required time:", standard)
     print()
